requirements.txt
openai-whisper
torch
torchaudio
ffmpeg-python
moviepy
pydub
tqdm
pinecone
google-generativeai
google-auth
python-dotenv
streamlit